requirements.txt
# linter formatter
mypy==1.3.0
black==23.12.1
flake8==6.0.0
isort==5.12.0
ruff==0.1.11

# test
pytest==7.3.2
syrupy==4.2.1
moto==4.2.0

# boto3
boto3==1.26.90
botocore==1.29.90
boto3-stubs[essential]==1.26.90

# cdk
aws-cdk-lib==2.118.0
constructs>=10.0.0,<11.0.0
cdk-nag==2.27.110

# powertools
aws-lambda-powertools[aws-sdk]==2.22.0

# aws mfa
aws-mfa==0.0.12

# batch
tqdm==4.66.1
